pyproject.toml
[project]
name = "netglance"
version = "0.1.0"
description = "Home network health checks — run by you or your AI"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9",
    "rich>=13.0",
    "scapy>=2.5",
    "dnspython>=2.4",
    "python3-nmap>=1.6",
    "zeroconf>=0.80",
    "icmplib>=3.0",
    "psutil>=5.9",
    "mac-vendor-lookup>=0.1.12",
    "pyyaml>=6.0",
    "httpx>=0.25",
    "plotext>=5.2",
    "fastmcp>=2.3",
    "fastapi>=0.110",
    "uvicorn>=0.27",
]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
]
docs = [
    "mkdocs>=1.6,<2",
    "mkdocs-shadcn>=0.9",
]

[project.scripts]
netglance = "netglance.cli:app"
netglance-mcp = "netglance.mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = ["netglance", ".venv", "dist", "build", ".git"]
